usersessions/pyproject.toml
[project]
name = "usersessions"
description = "TBD"
license = {text = "MIT" }
version = "1.0"
dependencies = [
	"nmrhubusage @ git+https://github.com/NMRbox/monosys.git@main#egg=nmrhubusage&subdirectory=nmrhubusage",
	"pyyaml",
	"systemhealthdata",
]
requires-python= ">= 3.10"
readme = "README.md"
authors = [
    {name = "Gerard"},
    {email = "gweatherby@uchc.edu"}
]

[tool.setuptools]
package-dir = {"" = "src" }  # Tells setuptools that the root package is in the "src" directory

[tool.setuptools.packages.find]
where = ["src"]  # Specify the "src" directory to look for packages

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

